pyproject.toml
[tool.poetry]
name = "hroshyk"
version = "0.3.0"
description = "Ukrainian donation platform"
authors = ["Gipss <gissa.dmtr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.6"
django-twitch-auth = "^0.1.0"
requests = "^2.28.2"
django-environ = "^0.9.0"
bcrypt = "^4.0.1"
django-debug-toolbar = "^3.8.1"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
psycopg2-binary = "^2.9.6"
django-filter = "^23.1"
django-cloudinary-storage = "^0.3.0"
Pillow = "^9.5.0"
django-cleanup = "^8.0.0"
django-jazzmin = "^2.6.0"
django-htmx = "^1.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
